Add minor fix
