+++ experiments/aws-ssm/aws-ssm-chaos-by-tag/test/test.yml
@@ -13,7 +13,7 @@ spec:
       labels:
         app: litmus-experiment
     spec:
-      serviceAccountName: ec2-terminate-by-id-sa
+      serviceAccountName: aws-ssm-chaos-by-tag-sa
       containers:
       - name: gotest
         image: busybox
@@ -24,7 +24,7 @@ spec:
           - name: LIB
             value: 'litmus'
 
-          - name: EC2_INSTANCE_ID
+          - name: EC2_INSTANCE_TAG
             value: ''
 
           - name: CHAOS_NAMESPACE
