+++ pkg/utils/common/common.go
@@ -9,9 +9,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/litmuschaos/litmus-go/pkg/clients"
 	"github.com/litmuschaos/litmus-go/pkg/events"
 	"github.com/litmuschaos/litmus-go/pkg/log"
@@ -166,24 +163,4 @@ func getEnvSource(apiVersion string, fieldPath string) apiv1.EnvVarSource {
 		},
 	}
 	return downwardENV
-}
-
-//GetAWSSession will return the aws session for a given region
-func GetAWSSession(region string) *session.Session {
-	return session.Must(session.NewSessionWithOptions(session.Options{
-		SharedConfigState: session.SharedConfigEnable,
-		Config:            aws.Config{Region: aws.String(region)},
-	}))
-}
-
-//CheckAWSError will return the aws errors
-func CheckAWSError(err error) error {
-	if aerr, ok := err.(awserr.Error); ok {
-		switch aerr.Code() {
-		default:
-			return errors.Errorf(aerr.Error())
-		}
-	} else {
-		return errors.Errorf(err.Error())
-	}
 }
